requirements.txt
fabric==1.5.1
yolk==0.4.1
cuisine==0.4.0
requests==0.9.3
python-cloudservers==1.2
